requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
asgiref==3.8.1
attrs==25.1.0
backoff==2.2.1
blinker==1.9.0
cartesia==1.3.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
dataclasses-json==0.6.7
deepgram-sdk==3.8.0
deprecation==2.1.0
distro==1.9.0
environs==14.1.1
fastapi==0.115.8
fastapi_cors==0.0.6
fixa-dev==0.0.4
Flask==3.0.2
flask-cors==5.0.1
flatbuffers==25.2.10
flexcache==0.3
flexparser==0.4
frozenlist==1.5.0
future==1.0.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
humanfriendly==10.0
idna==3.10
iterators==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
llvmlite==0.44.0
loguru==0.7.3
Markdown==3.7
MarkupSafe==3.0.2
marshmallow==3.26.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
ngrok==1.4.0
numba==0.61.0
numpy==1.26.4
onnxruntime==1.20.1
openai==1.59.9
packaging==24.2
pillow==11.1.0
Pint==0.24.4
pipecat==0.3.0
pipecat-ai==0.0.53
platformdirs==4.3.6
posthog==3.11.0
propcache==0.3.0
protobuf==5.29.3
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyloudnorm==0.1.1
python-dateutil==2.9.0.post0
python-deepcompare==2.1.0
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.3
resampy==0.4.3
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.3
tqdm==4.67.1
twilio==9.4.6
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==13.1
Werkzeug==3.1.3
yarl==1.18.3
